requirements.txt
playwright>=1.40.0
pyyaml>=6.0
pytest>=7.4.0